Use int16 for network params  (#193)
http://chess.grantnet.us/test/20362/

ELO   | 32.61 +- 11.90 (95%)
SPRT  | 8.0+0.08s Threads=1 Hash=8MB
LLR   | 3.00 (-2.94, 2.94) [0.00, 5.00]
GAMES | N: 1528 W: 427 L: 284 D: 817

Bench: 2168608

+++ src/network.h
@@ -32,9 +32,9 @@ public:
         hidden_neurons.back().fill(0.0f);
     }
 
-    int32_t calculate_last_layer();
+    int16_t calculate_last_layer();
 
-    int32_t feed(NetworkInput const &);
+    int16_t feed(NetworkInput const &);
 
     void update_hidden_layer(NetworkUpdateList const &);
 
@@ -55,11 +55,11 @@ public:
     }
 
 private:
-    std::vector<std::array<int32_t, HIDDEN_SIZE>> hidden_neurons;
+    std::vector<std::array<int16_t, HIDDEN_SIZE>> hidden_neurons;
 
-    static std::array<std::array<int32_t, HIDDEN_SIZE>, INPUT_SIZE> hidden_weights;
-    static std::array<int32_t, HIDDEN_SIZE> hidden_biases;
-    static std::array<int32_t, HIDDEN_SIZE> output_weights;
-    static int32_t output_bias;
+    static std::array<std::array<int16_t, HIDDEN_SIZE>, INPUT_SIZE> hidden_weights;
+    static std::array<int16_t, HIDDEN_SIZE> hidden_biases;
+    static std::array<int16_t, HIDDEN_SIZE> output_weights;
+    static int16_t output_bias;
     static std::uint32_t hash;
 };
